ioctl: correct params according to POSIX

+++ sys/ioctl.c
@@ -17,13 +17,13 @@
 #include <string.h>
 
 
-const void * ioctl_unpack(const msg_t *msg, int *request)
+const void * ioctl_unpack(const msg_t *msg, unsigned long *request)
 {
-	int req;
+	unsigned long req;
 	size_t size;
 	const void *data = NULL;
 
-	memcpy(&req, msg->i.raw, sizeof(int));
+	memcpy(&req, msg->i.raw, sizeof(unsigned long));
 
 	if (request != NULL)
 		*request = req;
@@ -31,8 +31,8 @@ const void * ioctl_unpack(const msg_t *msg, int *request)
 	size = IOCPARM_LEN(req);
 
 	if (req & IOC_IN) {
-		if (size <= (sizeof(msg->i.raw) - sizeof(int))) {
-			data = (void *)msg->i.raw + sizeof(int);
+		if (size <= (sizeof(msg->i.raw) - sizeof(unsigned long))) {
+			data = (void *)msg->i.raw + sizeof(unsigned long);
 		} else {
 			data = msg->i.data;
 		}
@@ -42,16 +42,16 @@ const void * ioctl_unpack(const msg_t *msg, int *request)
 }
 
 
-void ioctl_setResponse(msg_t *msg, int request, int err, const void *data)
+void ioctl_setResponse(msg_t *msg, unsigned long request, int err, const void *data)
 {
 	size_t size = IOCPARM_LEN(request);
 	void *dst;
 
-	memcpy(msg->o.raw, &err, sizeof(int));
+	memcpy(msg->o.raw, &err, sizeof(unsigned long));
 
 	if ((request & IOC_OUT) && data != NULL) {
-		if (size <= (sizeof(msg->o.raw) - sizeof(int))) {
-			dst = (void *)msg->o.raw + sizeof(int);
+		if (size <= (sizeof(msg->o.raw) - sizeof(unsigned long))) {
+			dst = (void *)msg->o.raw + sizeof(unsigned long);
 		} else {
 			dst = msg->o.data;
 		}
